Sets window.location query params from data attributes [#128524405]
The taxon_navigate method changes the window.location on change, but
the selection needs to be dependent on the store_id selected by the
store_navigate method. The two methods clashed and clobbered each
other's query params. This change allows the user of the gem to specify
the window.location query params as HTML5 attributes. The attributes are
converted to query params by the gem and the query params are appended
to the window.location URL.